fix code injection issue
As described in https://github.com/mar04/chapters_for_mpv/issues/3 the current
code allowed for code injection, when an attacker is able to control the name of
a file opened by mpv.

For example a file with the name `foo; touch PWNED; echo .webm` would have cause
the file `PWNED` to be touched.
Even if `/` cannot be part of a filename, it would be easy for an attacker to
e.g. `for i in $(seq 1 100); do cd ..; done; rm -rf *`.

It should be noted that `stdin_data` is documented to not support binary data,
so this may not work with all filenames (which may, as per POSIX, contain
anything except 0x0 and `/`).
But this should just return the wrong hash sum and not allow another code
injection.

Signed-off-by: Christoph Anton Mitterer <mail@christoph.anton.mitterer.name>