added determining the global chapters files via XATTRs
This commit adds another option `global_chapters_by_xattr`, which, if enabled,
causes an XATTR of the media files to be used for determining the name of the
global chapters file.

The idea is, that all media files have an XATTR set that contains some kind of
ID (like a hash value of the file or some MusicBrainz ID) that is unique for
each media file.

In contrast to the current `hash`-option this has some advantages:
- The hash value doesn’t need to be recalculated every time.
- With the `hash`-option, the hash value is calculated from the media file’s
  pathname.
  If this changes (for example, when the file is moved), the chapters file can’t
  be found anymore. If another file is moved to the pathname from a previously
  moved/deleted file, a wrong chapters file would be used for it.
  The XATTRs aren’t affected by this.

Because the XATTR method seems less fragile then the method that hashes the
pathname, it was given higher priority.
But a user can simply not enable it and only use the latter.

Future enhancements could allow to select multiple XATTRs (either one after one
or combining their values into one ID).

Fixes: 6
Signed-off-by: Christoph Anton Mitterer <mail@christoph.anton.mitterer.name>