manage this script as a “whole directory” mpv script
Previously, this script was a single file `chapters.lua`, to be placed in the
mpv script directory `~/.config/mpv/scripts`.

This commit changes this to use the mode, where a subdirectory of the mpv script
directory “belongs” to the whole script.

In that mode, the script’s name is determined by the subdirectory’s name.
And the script’s main file must be named `main.lua`.

The reason for this change is the 2nd part of this commit, which changes the
default location for the global directory to be the (dynamically determined)
script directory (i.e. the one that also contains `main.lua`).

This has the benefit, that we can write chapter files in that directory, without
fearing to clash with other scripts.
Since all chapter files get the extension `.ffmetadata`, there should be no
danger of accidentally having them interpreted by mvp or clash with our own
script file.

Signed-off-by: Christoph Anton Mitterer <mail@christoph.anton.mitterer.name>